+++ packages/react-native/src/components/PillGroup/Pill.props.ts
@@ -1,0 +1,13 @@
+import React from 'react';
+import { PressableProps } from 'react-native';
+
+export interface PillProps extends Omit<PressableProps, 'children'> {
+  /** Value returned when selected */
+  value: string;
+
+  /** Text label shown inside the pill */
+  label: string;
+
+  /** Left icon */
+  icon?: React.ComponentType<any>;
+}