+++ packages/react-native/src/components/index.ts
@@ -1,6 +1,7 @@
 // Custom
 export * from './Accordion';
 export * from './Alert';
+export * from './Avatar';
 export * from './Badge';
 export * from './Banner';
 export * from './BodyText';
@@ -13,6 +14,7 @@ export * from './Center';
 export * from './Checkbox';
 export * from './Container';
 export * from './CurrencyInput';
+export * from './DateInput';
 export * from './DatePicker';
 export * from './DatePickerInput';
 export * from './DescriptionList';
@@ -38,6 +40,7 @@ export * from './Link';
 export * from './List';
 export * from './Menu';
 export * from './Modal';
+export * from './PillGroup';
 export * from './ProgressStepper';
 export * from './Radio';
 export * from './RadioCard';
@@ -49,6 +52,7 @@ export * from './Switch';
 export * from './Tabs';
 export * from './Textarea';
 export * from './ThemedImage';
+export * from './Toast';
 export * from './ToggleButtonCard';
 export * from './VerificationInput';