+++ packages/react-native/src/utils/index.ts
@@ -3,6 +3,7 @@ export { default as formatThousands } from './formatThousands';
 export { default as getFlattenedColorValue } from './getFlattenedColorValue';
 export { default as getStyleValue } from './getStyleValue';
 export { default as hexWithOpacity } from './hexWithOpacity';
+export { getInitials } from './getInitials';
 export { default as isEqual } from './isEqual';
 export * from './styleUtils';
 export * from './themeValueHelpers';