+++ packages/react/src/components/List/ListItemLink.props.ts
@@ -1,0 +1,6 @@
+import { ComponentPropsWithout, RemovedProps } from '../../types/component-props';
+import { ListItemContentProps } from './ListItemContent.props';
+
+export interface ListItemLinkProps
+  extends ComponentPropsWithout<'a', RemovedProps>,
+    Omit<ListItemContentProps, keyof ComponentPropsWithout<'a', RemovedProps>> {}