+++ packages/react/src/index.ts
@@ -146,15 +146,18 @@ export type { TextAreaProps } from './components/TextArea/TextArea.props';
 
 export { List } from './components/List/List';
 export type { ListProps } from './components/List/List.props';
-
-export { ListItem } from './components/ListItem/ListItem';
-export type { ListItemProps } from './components/ListItem/ListItem.props';
-export { ListItemContent } from './components/ListItemContent/ListItemContent';
-export type { ListItemContentProps } from './components/ListItemContent/ListItemContent.props';
-export { ListItemButton } from './components/ListItemButton/ListItemButton';
-export type { ListItemButtonProps } from './components/ListItemButton/ListItemButton.props';
-export { ListItemLink } from './components/ListItemLink/ListItemLink';
-export type { ListItemLinkProps } from './components/ListItemLink/ListItemLink.props';
+export { ListItem } from './components/List/ListItem';
+export type { ListItemProps } from './components/List/ListItem.props';
+export { ListItemContent } from './components/List/ListItemContent';
+export type { ListItemContentProps } from './components/List/ListItemContent.props';
+export { ListItemButton } from './components/List/ListItemButton';
+export type { ListItemButtonProps } from './components/List/ListItemButton.props';
+export { ListItemLink } from './components/List/ListItemLink';
+export type { ListItemLinkProps } from './components/List/ListItemLink.props';
+export { ListActionButton } from './components/List/ListActionButton';
+export type { ListActionButtonProps } from './components/List/ListActionButton.props';
+export { ListActionLink } from './components/List/ListActionLink';
+export type { ListActionLinkProps } from './components/List/ListActionLink.props';
 
 export { Switch } from './components/Switch/Switch';
 export type { SwitchProps } from './components/Switch/Switch.props';
@@ -242,6 +245,16 @@ export type { ProgressStepButtonProps } from './components/ProgressStepper/Progr
 export { Avatar } from './components/Avatar/Avatar';
 export type { AvatarProps } from './components/Avatar/Avatar.props';
 
+export { Table } from './components/Table/Table';
+export type { TableProps } from './components/Table/Table.props';
+export { TableBody } from './components/TableBody/TableBody';
+export type { TableBodyProps } from './components/TableBody/TableBody.props';
+export { TableHeader } from './components/TableHeader/TableHeader';
+export { TableHeaderCell } from './components/TableHeaderCell/TableHeaderCell';
+export type { TableHeaderCellProps } from './components/TableHeaderCell/TableHeaderCell.props';
+export { TableRow } from './components/TableRow/TableRow';
+export { TableCell } from './components/TableCell/TableCell';
+
 export { Tabs } from './components/Tabs/Tabs';
 export { TabsList } from './components/Tabs/TabsList';
 export { Tab } from './components/Tabs/Tab';
@@ -250,6 +263,9 @@ export type { TabsProps } from './components/Tabs/Tabs.props';
 export type { TabsListProps } from './components/Tabs/TabsList.props';
 export type { TabProps } from './components/Tabs/Tab.props';
 export type { TabContentProps } from './components/Tabs/TabContent.props';
+
+export { Pagination } from './components/Pagination/Pagination';
+export type { PaginationProps } from './components/Pagination/Pagination.props';
 
 export { ProgressBar } from './components/ProgressBar/ProgressBar';
 export type { ProgressBarProps } from './components/ProgressBar/ProgressBar.props';